pyproject.toml
[project]
name = "jramonha-github-io"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "htmltools>=0.6.0",
    "ipykernel>=6.29.5",
    "nbclient>=0.10.2",
    "nbformat>=5.10.4",
    "pyyaml>=6.0.2",
]
